greekmarket-backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.32.3
Pillow==10.4.0

# Auth
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2

# Image upload
cloudinary==1.41.0

# Stripe payments
stripe==10.8.0
gunicorn==23.0.0
